+++ src/exceptiongroup/_catch.py
@@ -34,7 +34,16 @@ class _Catcher:
             elif unhandled is None:
                 return True
             else:
-                raise unhandled from None
+                if isinstance(exc, BaseExceptionGroup):
+                    try:
+                        raise unhandled from exc.__cause__
+                    except BaseExceptionGroup:
+                        # Change __context__ to __cause__ because Python 3.11 does this
+                        # too
+                        unhandled.__context__ = exc.__cause__
+                        raise
+
+                raise unhandled from exc
 
         return False
 
@@ -50,7 +59,12 @@ class _Catcher:
             matched, excgroup = excgroup.split(exc_types)
             if matched:
                 try:
-                    result = handler(matched)
+                    try:
+                        raise matched
+                    except BaseExceptionGroup:
+                        result = handler(matched)
+                except BaseExceptionGroup as new_exc:
+                    new_exceptions.extend(new_exc.exceptions)
                 except BaseException as new_exc:
                     new_exceptions.append(new_exc)
                 else:
@@ -66,9 +80,6 @@ class _Catcher:
         if new_exceptions:
             if len(new_exceptions) == 1:
                 return new_exceptions[0]
-
-            if excgroup:
-                new_exceptions.append(excgroup)
 
             return BaseExceptionGroup("", new_exceptions)
         elif (