+++ tests/test_catch.py
@@ -148,9 +148,41 @@ def test_catch_handler_raises():
     def handler(exc):
         raise RuntimeError("new")
 
-    with pytest.raises(RuntimeError, match="new"):
+    with pytest.raises(RuntimeError, match="new") as exc:
         with catch({(ValueError, ValueError): handler}):
-            raise ExceptionGroup("booboo", [ValueError("bar")])
+            excgrp = ExceptionGroup("booboo", [ValueError("bar")])
+            raise excgrp
+
+    context = exc.value.__context__
+    assert isinstance(context, ExceptionGroup)
+    assert str(context) == "booboo (1 sub-exception)"
+    assert len(context.exceptions) == 1
+    assert isinstance(context.exceptions[0], ValueError)
+    assert exc.value.__cause__ is None
+
+
+def test_bare_raise_in_handler():
+    """Test that a bare "raise"  "middle" ecxeption group gets discarded."""
+
+    def handler(exc):
+        raise
+
+    with pytest.raises(ExceptionGroup) as excgrp:
+        with catch({(ValueError,): handler, (RuntimeError,): lambda eg: None}):
+            try:
+                first_exc = RuntimeError("first")
+                raise first_exc
+            except RuntimeError as exc:
+                middle_exc = ExceptionGroup(
+                    "bad", [ValueError(), ValueError(), TypeError()]
+                )
+                raise middle_exc from exc
+
+    assert len(excgrp.value.exceptions) == 2
+    assert all(isinstance(exc, ValueError) for exc in excgrp.value.exceptions)
+    assert excgrp.value is not middle_exc
+    assert excgrp.value.__cause__ is first_exc
+    assert excgrp.value.__context__ is first_exc
 
 
 def test_catch_subclass():