+++ tests/test_catch_py311.py
@@ -128,11 +128,19 @@ def test_catch_full_match():
     reason="Behavior was changed in 3.11.4",
 )
 def test_catch_handler_raises():
-    with pytest.raises(RuntimeError, match="new"):
+    with pytest.raises(RuntimeError, match="new") as exc:
         try:
-            raise ExceptionGroup("booboo", [ValueError("bar")])
+            excgrp = ExceptionGroup("booboo", [ValueError("bar")])
+            raise excgrp
         except* ValueError:
             raise RuntimeError("new")
+
+    context = exc.value.__context__
+    assert isinstance(context, ExceptionGroup)
+    assert str(context) == "booboo (1 sub-exception)"
+    assert len(context.exceptions) == 1
+    assert isinstance(context.exceptions[0], ValueError)
+    assert exc.value.__cause__ is None
 
 
 def test_catch_subclass():
@@ -146,3 +154,25 @@ def test_catch_subclass():
     assert isinstance(lookup_errors[0], ExceptionGroup)
     exceptions = lookup_errors[0].exceptions
     assert isinstance(exceptions[0], KeyError)
+
+
+def test_bare_raise_in_handler():
+    """Test that the "middle" ecxeption group gets discarded."""
+    with pytest.raises(ExceptionGroup) as excgrp:
+        try:
+            try:
+                first_exc = RuntimeError("first")
+                raise first_exc
+            except RuntimeError as exc:
+                middle_exc = ExceptionGroup(
+                    "bad", [ValueError(), ValueError(), TypeError()]
+                )
+                raise middle_exc from exc
+        except* ValueError:
+            raise
+        except* TypeError:
+            pass
+
+    assert excgrp.value is not middle_exc
+    assert excgrp.value.__cause__ is first_exc
+    assert excgrp.value.__context__ is first_exc