+++ CHANGES.rst
@@ -7,6 +7,10 @@ This library adheres to `Semantic Versioning 2.0 <http://semver.org/>`_.
 - `catch()` now raises a `TypeError` if passed an async exception handler instead of
   just giving a `RuntimeWarning` about the coroutine never being awaited. (#66, PR by
   John Litborn)
+- Fixed plain ``raise`` statement in an exception handler callback to work like a
+  ``raise`` in an ``except*`` block
+- Fixed new exception group not being chained to the original exception when raising an
+  exception group from exceptions raised in handler callbacks
 
 **1.1.2**